added readme and refactored files
