--- Montador/MONTADOR.cpp
@@ -1,11 +1,0 @@
-#include <bits/stdc++.h>
-using namespace std;
-
-int main()
-{
-    ios_base::sync_with_stdio(0);
-    cin.tie(0);
-    cout.tie(0);
-
-    return 0;
-}
